lt-qcom-openembedded: Adapt submit_for_testing with LKFT-like approach

+++ lt-qcom-openembedded/submit_for_testing.sh
@@ -1,10 +1,15 @@
 #!/bin/bash
 
-rm -rf configs
-git clone --depth 1 http://git.linaro.org/ci/job/configs.git
+if [ -z "${DRY_RUN}" ]; then
+  rm -rf configs
+  git clone --depth 1 http://git.linaro.org/ci/job/configs.git
+  export CONFIG_PATH=$(realpath configs)
 
-# Install jinja2-cli and ruamel.yaml, required by submit_for_testing.py
-pip install --user --force-reinstall jinja2-cli ruamel.yaml
+  # Install jinja2-cli and ruamel.yaml, required by submit_for_testing.py
+  pip install --user --force-reinstall jinja2-cli ruamel.yaml
+else
+  export CONFIG_PATH=$(realpath ../)
+fi
 
 # main parameters
 export DEPLOY_OS=oe
@@ -16,6 +21,7 @@ export BOOT_URL_COMP=
 export LXC_BOOT_FILE=$(basename ${BOOT_URL})
 
 export RESIZE_ROOTFS=
+export LAVA_JOB_PRIORITY="medium"
 
 case "${MACHINE}" in
   dragonboard-410c|dragonboard-820c)
@@ -52,7 +58,10 @@ case "${MACHINE}" in
         export ROOTFS_URL=${ROOTFS_SPARSE_BUILD_URL}
         export ROOTFS_URL_COMP="gz"
         export LXC_ROOTFS_FILE=$(basename ${ROOTFS_URL} .gz)
-        python configs/openembedded-lkft/submit_for_testing.py \
+
+	LAVA_TEMPLATE_PATH=${CONFIG_PATH}/lt-qcom/lava-job-definitions
+	cd ${LAVA_TEMPLATE_PATH}
+        python ${CONFIG_PATH}/openembedded-lkft/submit_for_testing.py \
             --device-type ${DEVICE_TYPE} \
             --build-number ${BUILD_NUMBER} \
             --lava-server ${LAVA_SERVER} \
@@ -61,9 +70,10 @@ case "${MACHINE}" in
             --qa-server-project openembedded-rpb-${MANIFEST_BRANCH} \
             --env-suffix="-${DISTRO}" \
             --git-commit ${BUILD_NUMBER} \
-            --template-path configs/lt-qcom/lava-job-definitions \
-            --template-base-pre base_template.yaml \
-            --template-names template.yaml template-wifi.yaml template-bt.yaml
+	    --template-path "${LAVA_TEMPLATE_PATH}" \
+	    --testplan-path "${LAVA_TEMPLATE_PATH}" \
+	    ${DRY_RUN} \
+            --test-plan testplan/main.yaml testplan/wifi.yaml testplan/bt.yaml
       ;;
       rpb-wayland)
         echo "Currently no tests for rpb-wayland"
